requirements.txt
colorama
pyyaml
rich